requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
transformers
sentence-transformers
accelerate
huggingface_hub
chromadb
pypdf
pymupdf
python-dotenv
fastapi
flask
uvicorn
gunicorn
requests
httpx
